Erased redundant comments, moving context to App.js, minor styling changes
