+++ src/components/BoardList/BoardList.component.js
@@ -6,25 +6,7 @@ import CreateBoardInput from "../CreateBoardInput/CreateBoardInput.component";
 import { BoardContext } from "../../contexts/BoardData.context";
 import { useContext } from "react";
 
-// import { getBoardStorage } from "../../assets/utils";
-
-// import { useState, useEffect } from "react";
-
 const BoardList = () => {
-    // const [boards, setBoards] = useState([]);
-
-    // useEffect(() => {
-    //     try {
-    //         const getBoardData = async () => {
-    //             const res = await JSON.parse(getBoardStorage());
-    //             setBoards(res);
-    //         };
-    //         getBoardData();
-    //     } catch (error) {
-    //         console.log(error);
-    //     }
-    // }, []);
-
     const { boards, setBoards } = useContext(BoardContext);
 
     return (
