+++ src/components/BoardList/BoardList.styles.scss
@@ -7,7 +7,7 @@
         align-items: center;
         justify-content: space-between;
         margin-left: 50px;
-        width: 99%;
+        width: 98.5%;
     }
 
     .board-list {
