+++ src/components/CreateBoardInput/CreateBoardInput.styles.scss
@@ -1,55 +1,60 @@
 @import "../../assets/style-elements.scss";
 
-.form {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: center;
-}
-
-.input-form {
+.container {
     position: relative;
-    width: 250px;
-    height: 38px;
-    padding: 12px 24px;
-    border: 1px solid #c7d4d9;
-    border-radius: $border-radius;
-    font-size: 16px;
-    transition: all 500ms;
 
-    &:focus {
-        outline: none;
+    .form {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        justify-content: center;
     }
-}
 
-.button-style {
-    background-color: $main-color;
-    font-weight: bold;
-    color: white;
-    padding: 8px 15px;
-    border: none;
-    border-radius: 7px;
-    box-shadow: $box-shadow;
-    transition: all 50ms;
-    cursor: pointer;
-    transform: scale(0.9);
+    .input-form {
+        width: 250px;
+        height: 38px;
+        padding: 12px 24px;
+        border: 1px solid #c7d4d9;
+        border-radius: $border-radius;
+        font-size: 16px;
+        transition: all 500ms;
 
-    &.cancel-btn {
-        background-color: white;
-        color: $main-color;
-        box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
+        &:focus {
+            outline: none;
+        }
     }
-}
 
-.success {
-    color: green;
-    font-size: 20px;
-    font-weight: bold;
-    margin: 10px 8px;
-}
+    .button-style {
+        background-color: $main-color;
+        font-weight: bold;
+        color: white;
+        padding: 8px 15px;
+        border: none;
+        border-radius: 7px;
+        box-shadow: $box-shadow;
+        transition: all 50ms;
+        cursor: pointer;
+        transform: scale(0.9);
 
-.error {
-    position: absolute;
-    font-size: 12px;
-    color: red;
+        &.cancel-btn {
+            background-color: white;
+            color: $main-color;
+            box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
+        }
+    }
+
+    .success {
+        color: green;
+        font-size: 20px;
+        font-weight: bold;
+        position: absolute;
+        left: 24px;
+    }
+
+    .error {
+        position: absolute;
+        left: 24px;
+        font-size: 12px;
+        color: red;
+    }
 }
